bot_service/requirements.txt
python-telegram-bot==22.1
google-generativeai==0.8.5
requests==2.32.3
python-dotenv==1.0.1 